refactor(rules-builder): extract ItemRulesTab state to factory
Move the reactive state and JSON-editor handlers in ItemRulesTab into a
co-located .svelte.ts factory. openBuilder stays in the component since
it imports a Svelte component (RulesBuilderWindow) which can't be
referenced from a plain .ts module.
